[MOVE] 하이라이팅관련 ViewModifier 코드 이동
