+++ Sofa/Configuration/Extensions/Ex+View.swift
@@ -107,28 +107,3 @@ struct TextFieldModifier: ViewModifier {
       )
   }
 }
-
-// TextField 선택 시 하이라이트 기능
-struct RectangleHighlightedViewModifier: ViewModifier {
-  var firstLineWidth: CGFloat
-  var secondLineWidth: CGFloat
-  
-  func body(content: Content) -> some View {
-    content
-      .overlay(RoundedRectangle(cornerRadius: 6).strokeBorder(Color(hex: "#43A047"), lineWidth: firstLineWidth))
-      .padding(4)
-      .overlay(RoundedRectangle(cornerRadius: 10).strokeBorder(Color(hex:"4CAF50").opacity(0.5), lineWidth: secondLineWidth))
-  }
-}
-
-// Circle 선택 시 하이라이트 기능
-struct CircleHighlightedViewModifier: ViewModifier {
-  var lineWidth: CGFloat
-  
-  func body(content: Content) -> some View {
-    content
-      .padding(4)
-      .overlay(Circle().strokeBorder(Color(hex:"4CAF50").opacity(0.5), lineWidth: lineWidth))
-  }
-}
-
